crawler/requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
tldextract==5.1.2
supabase==2.4.6
python-dotenv==1.0.1
lxml==5.2.1
fastapi==0.115.6
uvicorn==0.32.1
